+++ feature/mypage/src/main/java/com/record/mypage/navigation/MypageNavigation.kt
@@ -29,7 +29,7 @@ fun NavGraphBuilder.mypageNavGraph(
     navigateToSetting: () -> Unit,
     navigateToFollowing: () -> Unit,
     navigateToFollower: () -> Unit,
-    navigateToVideo: (VideoType, Int) -> Unit,
+    navigateToVideo: (VideoType, Long) -> Unit,
 ) {
     composable(route = MypageRoute.route) {
         MypageRoute(
